Managed to get the searchable to work.
Modified the BundleDecoder to be more descriptive.

Modified the static example data inside Character and Quest. Not sure about the performance cost of it since it needs to load the json and use the first element of it as an example.

Added prompts to the .searchable text fields.

Changed the 1 main NavigationStack in the ContentView to be multiple NavigationStacks inside the TabView. This fixed the swift quirk with the .searchables not working properly.

Added Sections to the QuestDetailView.